Feature/축제 상세보기 기능 개발

+++ src/main/java/com/example/tripKo/domain/place/dao/PlaceFestivalJPARepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PlaceFestivalJPARepository extends JpaRepository<PlaceFestival, Integer> {
+public interface PlaceFestivalJPARepository extends JpaRepository<PlaceFestival, Long> {
     @Query("SELECT p FROM PlaceFestival p WHERE UPPER(p.place.address.addressCategory.sidoName) = UPPER(:location)")
     List<PlaceFestival> findAllByLocation(@Param("location") String location);
